requirements.txt
uvicorn==0.21.0
fastapi==0.92.0
requests==2.28.2
python-dotenv==1.0.0